refactor: release note automation
* Disable authentication. No need for authentication to download public data from S3
* reduce complexity in kernelurl extraction
* Let python directly do the API Call
